Feature/audit logs

+++ IoTPlatform.API/Controllers/ClientAttributeController.cs
@@ -1,4 +1,5 @@
 ﻿using IoTPlatform.Domain.Models;
+using IoTPlatform.Domain.Models.AuditLog;
 using IoTPlatform.Infrastructure.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
@@ -9,10 +10,14 @@ namespace IoTPlatform.API.Controllers
     public class ClientAttributeController : ControllerBase
     {
         private readonly IClientAttributeService _clientAttributeService;
+        private readonly IUserService _userService;
+        private readonly IAuditLogService _auditLogService;
 
-        public ClientAttributeController(IClientAttributeService clientAttributeService)
+        public ClientAttributeController(IClientAttributeService clientAttributeService, IUserService userService, IAuditLogService auditLogService)
         {
             _clientAttributeService = clientAttributeService;
+            _userService = userService;
+            _auditLogService = auditLogService;
         }
 
         [HttpGet]
@@ -41,12 +46,25 @@ namespace IoTPlatform.API.Controllers
         public async Task<ActionResult> UpdateClientAttribute(string id, ClientAttribute clientAttribute)
         {
             var result = await _clientAttributeService.UpdateClientAttributeAsync(id, clientAttribute);
+
+            var userInfor = _userService.GetUserInformation();
+            await _auditLogService.AddAnAuditLogAsync(DateTime.Now, EntityType.ClientAttribute, result.AttributeID, "", userInfor[0], userInfor[1], ActionType.Create);
+
             return new JsonResult(new { result });
         }
 
         [HttpDelete("{id}")]
         public async Task<ActionResult> RemoveClientAttribute(string id)
         {
+            var removeClientAttribute = await _clientAttributeService.FindClientAttributeByIdAsync(id);
+            if (removeClientAttribute == null)
+            {
+                return NotFound();
+            }
+
+            var userInfor = _userService.GetUserInformation();
+            await _auditLogService.AddAnAuditLogAsync(DateTime.Now, EntityType.ClientAttribute, removeClientAttribute.AttributeID, "", userInfor[0], userInfor[1], ActionType.Delete);
+            
             var result = await _clientAttributeService.RemoveClientAttributeAsync(id);
             return new JsonResult(new { result });
         }
