+++ IoTPlatform.API/Controllers/RuleChainController.cs
@@ -1,4 +1,5 @@
 ﻿using IoTPlatform.Domain.Models;
+using IoTPlatform.Domain.Models.AuditLog;
 using IoTPlatform.Infrastructure.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
@@ -9,16 +10,25 @@ namespace IoTPlatform.API.Controllers
     public class RuleChainController : ControllerBase
     {
         private readonly IRuleChainService _ruleChainService;
+        private readonly IUserService _userService;
+        private readonly IAuditLogService _auditLogService;
 
-        public RuleChainController(IRuleChainService ruleChainService)
+        public RuleChainController(IRuleChainService ruleChainService, IUserService userService, IAuditLogService auditLogService)
         {
             _ruleChainService = ruleChainService;
+            _userService = userService;
+            _auditLogService = auditLogService;
         }
 
         [HttpPost]
         public async Task<ActionResult> AddRuleChain(RuleChain ruleChain)
         {
             var result = await _ruleChainService.AddRuleChainAsync(ruleChain);
+
+            var userInfor = _userService.GetUserInformation();
+            await _auditLogService.AddAnAuditLogAsync(DateTime.Now, EntityType.RuleChain, result.RuleChainID,
+                result.RuleChainName, userInfor[0], userInfor[1], ActionType.Create);
+
             return new JsonResult(new { result });
         }
 
@@ -48,12 +58,27 @@ namespace IoTPlatform.API.Controllers
         public async Task<ActionResult> UpdateRuleChain(string id, RuleChain ruleChain)
         {
             var result = await _ruleChainService.UpdateRuleChainAsync(id, ruleChain);
+
+            var userInfor = _userService.GetUserInformation();
+            await _auditLogService.AddAnAuditLogAsync(DateTime.Now, EntityType.RuleChain, result.RuleChainID,
+                result.RuleChainName, userInfor[0], userInfor[1], ActionType.Update);
+
             return new JsonResult(new { result });
         }
 
         [HttpDelete("{id}")]
         public async Task<ActionResult> RemoveRuleChain(string id)
-        {
+        { 
+            var removeRuleChain = await _ruleChainService.FindRuleChainByIdAsync(id);
+            if (removeRuleChain == null)
+            {
+                return NotFound();
+            }
+
+            var userInfor = _userService.GetUserInformation();
+            await _auditLogService.AddAnAuditLogAsync(DateTime.Now, EntityType.RuleChain, removeRuleChain.RuleChainID,
+                removeRuleChain.RuleChainName, userInfor[0], userInfor[1], ActionType.Delete);
+           
             var result = await _ruleChainService.RemoveRuleChainAsync(id);
             return new JsonResult(new { result });
         }
